Remove dtype argument from hk.Linear initializers to fix TypeError

+++ VishwamAI/scripts/model_architecture.py
@@ -17,8 +17,8 @@ class VishwamAIModel(hk.Module):
             lambda x: hk.Sequential([
                 hk.Embed(vocab_size=20000, embed_dim=128, w_init=hk.initializers.VarianceScaling(1.0, "fan_avg")),
                 lambda x: self.attention(x, x, x),  # Keep inputs as integers for embedding
-                hk.Linear(512, w_init=hk.initializers.VarianceScaling(1.0, "fan_avg"), dtype=np.float32),
-                hk.Linear(256, w_init=hk.initializers.VarianceScaling(1.0, "fan_avg"), dtype=np.float32)
+                hk.Linear(512, w_init=hk.initializers.VarianceScaling(1.0, "fan_avg")),
+                hk.Linear(256, w_init=hk.initializers.VarianceScaling(1.0, "fan_avg"))
             ])(x),
             apply_rng=True
         )
@@ -35,8 +35,8 @@ class VishwamAIModel(hk.Module):
             lambda x: hk.Sequential([
                 hk.Embed(vocab_size=10000, embed_dim=64, w_init=hk.initializers.VarianceScaling(1.0, "fan_avg")),
                 lambda x: self.attention(x, x, x),  # Keep inputs as integers for embedding
-                hk.Linear(256, w_init=hk.initializers.VarianceScaling(1.0, "fan_avg"), dtype=np.float32),
-                hk.Linear(128, w_init=hk.initializers.VarianceScaling(1.0, "fan_avg"), dtype=np.float32)
+                hk.Linear(256, w_init=hk.initializers.VarianceScaling(1.0, "fan_avg")),
+                hk.Linear(128, w_init=hk.initializers.VarianceScaling(1.0, "fan_avg"))
             ])(x),
             apply_rng=True
         ) for _ in range(self.num_experts)]
